+++ src/main/java/com/vonage/client/verify2/FragmentChannel.java
@@ -1,0 +1,47 @@
+/*
+ *   Copyright 2024 Vonage
+ *
+ *   Licensed under the Apache License, Version 2.0 (the "License");
+ *   you may not use this file except in compliance with the License.
+ *   You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *   Unless required by applicable law or agreed to in writing, software
+ *   distributed under the License is distributed on an "AS IS" BASIS,
+ *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *   See the License for the specific language governing permissions and
+ *   limitations under the License.
+ */
+package com.vonage.client.verify2;
+
+import com.fasterxml.jackson.annotation.JsonCreator;
+import com.fasterxml.jackson.annotation.JsonValue;
+
+/**
+ * Represents the template channel type in a {@link TemplateFragment}.
+ */
+public enum FragmentChannel {
+	SMS,
+	VOICE;
+
+	@JsonValue
+	@Override
+	public String toString() {
+		return name().toLowerCase();
+	}
+
+	/**
+	 * Parse a channel from a string.
+	 *
+	 * @param value The string to parse.
+	 *
+	 * @return The parsed channel, or {@code null} if the string is null.
+	 * @throws IllegalArgumentException If the string does not match a known channel.
+	 */
+	@JsonCreator
+	public static FragmentChannel fromString(String value) {
+		if (value == null) return null;
+		return valueOf(value.toUpperCase());
+	}
+}