+++ src/main/java/com/vonage/client/verify2/ListTemplatesRequest.java
@@ -1,0 +1,33 @@
+/*
+ *   Copyright 2024 Vonage
+ *
+ *   Licensed under the Apache License, Version 2.0 (the "License");
+ *   you may not use this file except in compliance with the License.
+ *   You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *   Unless required by applicable law or agreed to in writing, software
+ *   distributed under the License is distributed on an "AS IS" BASIS,
+ *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *   See the License for the specific language governing permissions and
+ *   limitations under the License.
+ */
+package com.vonage.client.verify2;
+
+import com.vonage.client.common.HalFilterRequest;
+import java.util.UUID;
+
+/**
+ * Represents a filter request for listing templates and fragments.
+ *
+ * @since 8.13.0
+ */
+final class ListTemplatesRequest extends HalFilterRequest {
+	final UUID templateId;
+
+	ListTemplatesRequest(Integer page, Integer pageSize, UUID templateId) {
+		super(page, pageSize, null);
+		this.templateId = templateId;
+	}
+}