+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+# [8.13.0] - 2024-1?-??
+- Added support for Verify Templates
+
 # [8.12.0] - 2024-10-21
 - Added `network_apis` capability to Application API
 - Added `mode` property to `InputAction` NCCO